Input Validation Added
Added a new exception for when input validation fails called DoubleInputValidationException. Right now the only input we take fropm the user is double anyway, but i figured if we add more later and have multiple types of data we need to validate it would be best to have it classified names this way. I also created a new class called Validation Which has a static method ValidateInput which takes in the input in string form and a regex string for it to test against. It returns true if it matches the input to the regex and false if there is no match. I uses this method in CheckingAccount and SavingsAccount in the Deposit and withdrawl methods. If it returns false, the those methods will now throw a DoubleInputvalidationException. I added an extra catch statement to all the tests to handle this new exception and I added a JUnit test to test the ValidateInput method.